refactor(*): move database tables to models directory
Establishing a new directory/package layout.
