--- src/ramstk/data/storage/__init__.py
@@ -1,10 +1,0 @@
-# pylint: disable=unused-import
-# -*- coding: utf-8 -*-
-#
-#       ramstk.data.storage.__init__.py is part of The RAMSTK Project
-#
-# All rights reserved.
-# Copyright 2007 - 2017 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
-"""The RAMSTK Data storage package."""
-# RAMSTK Local Imports
-from .programdb import RAMSTKAllocation, RAMSTKSimilarItem
