+++ src/ramstk/models/__init__.py
@@ -1,8 +1,8 @@
 # pylint: disable=unused-import
 # -*- coding: utf-8 -*-
 #
-#       ramstk.data.__init__.py is part of The RAMSTK Project
+#       ramstk.models.__init__.py is part of The RAMSTK Project
 #
 # All rights reserved.
 # Copyright 2007 - 2017 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
-"""The RAMSTK Data topic package."""
+"""The RAMSTK database models package."""
