+++ src/ramstk/models/programdb/__init__.py
@@ -1,11 +1,11 @@
 # pylint: disable=unused-import
 # -*- coding: utf-8 -*-
 #
-#       ramstk.data.storage.programdb.__init__.py is part of The RAMSTK Project
+#       ramstk.models.programdb.__init__.py is part of The RAMSTK Project
 #
 # All rights reserved.
 # Copyright 2007 - 2017 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
-"""The RAMSTK program database package."""
+"""The RAMSTK program database models package."""
 
 # RAMSTK Local Imports
 from .RAMSTKAllocation import RAMSTKAllocation
