+++ tests/models/programdb/test_ramstkdesignelectric.py
@@ -10,9 +10,7 @@
 import pytest
 
 # RAMSTK Package Imports
-from ramstk.data.storage.programdb.RAMSTKDesignElectric import (
-    RAMSTKDesignElectric
-)
+from ramstk.models.programdb.RAMSTKDesignElectric import RAMSTKDesignElectric
 
 ATTRIBUTES = {
     'voltage_ac_operating': 0.0,
